Moves PromiseArray to separate file
